+++ src/main/java/domain/game/Command.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import java.util.Arrays;
