+++ src/test/java/domain/FixtureCard.java
@@ -1,10 +1,14 @@
 package domain;
 
+import domain.card.Card;
+import domain.card.Rank;
+import domain.card.Suit;
+
 public class FixtureCard {
-    static final Card TWO_HEARTS = Card.of(Rank.TWO, Suit.HEARTS);
-    static final Card SIX_HEARTS = Card.of(Rank.SIX, Suit.HEARTS);
-    static final Card SEVEN_HEARTS = Card.of(Rank.SEVEN, Suit.HEARTS);
-    static final Card TEN_HEARTS = Card.of(Rank.TEN, Suit.HEARTS);
-    static final Card ACE_HEARTS = Card.of(Rank.ACE, Suit.HEARTS);
+    public static final Card TWO_HEARTS = Card.of(Rank.TWO, Suit.HEARTS);
+    public static final Card SIX_HEARTS = Card.of(Rank.SIX, Suit.HEARTS);
+    public static final Card SEVEN_HEARTS = Card.of(Rank.SEVEN, Suit.HEARTS);
+    public static final Card TEN_HEARTS = Card.of(Rank.TEN, Suit.HEARTS);
+    public static final Card ACE_HEARTS = Card.of(Rank.ACE, Suit.HEARTS);
 }
