+++ src/test/java/domain/card/CardDeckTest.java
@@ -1,9 +1,9 @@
-package domain;
+package domain.card;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 class CardDeckTest {
